# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Repair Store",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.nethanelfitoussi.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "List Clients",
                "responses": {
                    "200": {
                        "description": "Success: List all Clients",
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/Clients": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create Client",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Client",
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get Clients by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified client by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Client",
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update Client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified client by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Client updated",
                        "schema": {
                            "$ref": "#/definitions/Client"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete client",
                "description": "Delete the specified client by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
        "/api/produits": {
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Create produit",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Produit"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Produit"
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Client": {
            "required": [
                "name",
                "phone_number"
            ],
            "properties": {
                "name": {
                    "description": "Nom du client",
                    "type": "string",
                    "example": "Nethanel Fitoussi"
                },
                "phone_number": {
                    "description": "Nom du client",
                    "type": "string",
                    "example": "Nethanel Fitoussi"
                },
                "email": {
                    "description": "Email du client",
                    "type": "string",
                    "example": "contact@nethanelfitoussi.com"
                },
                "dept": {
                    "description": "Argent du par le client",
                    "type": "string",
                    "example": "430 euros"
                },
                "loan": {
                    "description": "Si l'on a pr\u00e9ter un t\u00e9l\u00e9phone",
                    "type": "string",
                    "example": "Iphone 4 blanc + clable"
                }
            }
        },
        "Comments Intervention": {
            "required": [],
            "properties": {
                "comments": {
                    "description": "Commentaire",
                    "type": "text",
                    "example": "Iphone acheter a Albert pour 500 ils Puis d\u00e9bloquer par nos soin"
                },
                "Name": {
                    "description": "Commentaire",
                    "type": "text",
                    "example": "Iphone acheter a Albert pour 500 ils Puis d\u00e9bloquer par nos soin"
                }
            }
        },
        "Itervention": {
            "required": [
                "password",
                "price"
            ],
            "properties": {
                "Password": {
                    "description": "Mot de passe du telephone",
                    "type": "string",
                    "example": "*********"
                },
                "price": {
                    "description": "Prix de l'intervention",
                    "type": "string",
                    "example": "39\u20ac"
                }
            }
        },
        "Produit": {
            "required": [
                "brand",
                "model",
                "price"
            ],
            "properties": {
                "reference": {
                    "description": "Mot de passe du telephone",
                    "type": "string",
                    "example": "IpXsMax128"
                },
                "brand": {
                    "description": "Marque du produit",
                    "type": "string",
                    "example": "Apple"
                },
                "model": {
                    "description": "Model du produit",
                    "type": "string",
                    "example": "Iphone Xs Max"
                },
                "price": {
                    "description": "Prix de vente",
                    "type": "string",
                    "example": "300\u20ac"
                },
                "buying_price": {
                    "description": "Prix a l'achat du produit",
                    "type": "string",
                    "example": "100\u20ac"
                },
                "stock": {
                    "description": "How many pieces in stock",
                    "type": "integer",
                    "example": "10"
                },
                "visible": {
                    "description": "If the product is visible or not",
                    "type": "bolean",
                    "example": "1 or 0"
                }
            }
        },
        "Status": {
            "required": [
                "types"
            ],
            "properties": {
                "types": {
                    "description": "Quel status de l'intervention",
                    "type": "string",
                    "example": "Terminer, en cour, payer, en attente, command\u00e9..."
                }
            }
        },
        "Product type": {
            "required": [
                "types"
            ],
            "properties": {
                "types": {
                    "description": "Type de produit",
                    "type": "string",
                    "example": "Computer, phone ...s"
                }
            }
        }
    },
    "securityDefinitions": {
        "access_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}